move schema test
